requirements.txt
python-telegram-bot==21.9
yt-dlp
python-dotenv
motor
